requirements.txt
asgiref==3.3.1
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.3
gunicorn==20.0.4
h5py==3.1.0
Keras==2.4.3
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0

djangorestframework~=3.12.2